+++ src/api/transaction/transactionAsyncApi.ts
@@ -31,7 +31,12 @@ import {
 
 const ENDPOINT = 'transactions';
 
-type ListResponse = { transactions: TransactionAsync[] };
+/**
+ * @param next Is a continuation token that can be used to fetch the next page.
+ *   `next` value is undefined if there are no more results to fetch.
+ */
+type ListResponse = { transactions: TransactionAsync[]; next?: string };
+
 type SingleResponse = { transaction: TransactionAsync };
 type CancelResponse = {
   message?: string;
@@ -54,10 +59,22 @@ export class TransactionAsyncApi extends Base {
     };
   }
 
-  async listTransactions(options?: TransactionListOptions) {
-    const result = await this.get<ListResponse>(ENDPOINT, options);
+  async listTransactions(
+    options?: TransactionListOptions,
+  ): Promise<ListResponse> {
+    const { sortBy, ...opts } = options ?? ({} as any);
 
-    return result.transactions;
+    Object.keys(opts).forEach(key => {
+      if (opts[key] instanceof Date) {
+        opts[key] = opts[key].getTime();
+      }
+    });
+
+    if (sortBy) {
+      opts.$sortby = `${sortBy.order === 'desc' ? '-' : ''}${sortBy.field}`;
+    }
+
+    return this.get<ListResponse>(ENDPOINT, opts);
   }
 
   async getTransaction(transactionId: string) {